add a new blob test case
